// src/app/models/bl_pda.model.ts
/* eslint-disable @typescript-eslint/naming-convention */
export class BLLPDA{
	DocType?: string;
	BLType?: string;
	Date?: Date;
	BLNum?: string;
	MontantHT?: number;
	TVA?: number;
	Remise?: number;
	Timbre?: number;
	MontantTTC?: number;
	CodeClient?: string;
	IdCamionPDA?: string;
	IdPromotion?: number;
	SommmePoint?: number;
	Point?: number;
}